+++ .env
@@ -1,3 +1,4 @@
 export DATASET_ROOT="${HOME}/datasets/ModaNet"
 export PIPENV_PIPFILE="${DATASET_ROOT}/Pipfile"
 export PAPERDOLL_JSONFILE="${HOME}/datasets/PaperDoll/raw/labels/paperdoll.json"
+export CHICTOPIA_SQLITE="${HOME}/datasets/PaperDoll/raw/chictopia.sqlite3"
