+++ .gitignore
@@ -6,3 +6,4 @@ raw/labels/*
 !raw/labels/categories.tsv
 !raw/labels/mean.txt
 main/*
+tiny/*
