+++ Pipfile
@@ -23,6 +23,7 @@ Cython = "*"
 pycocotools = "*"
 mljsp = {git = "https://github.com/hrsma2i/ml_json_processor.git"}
 matplotlib = "*"
+aiodl = {git = "https://github.com/hrsma2i/aiodl.git"}
 
 [requires]
 python_version = "3.6"
